# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Forum Furiosum API
  version: 1.0.0
  description: REST API for Forum Furiosum web forum
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ff_session
...
paths:
...
  /threads/{id}:
    get:
      summary: Get thread
      tags: [Threads]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thread object
        '404':
          description: Not found
    patch:
      summary: Edit thread
      tags: [Threads]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 120
                body:
                  type: string
                  minLength: 1
                  maxLength: 10000
                reason:
                  type: string
      responses:
        '200':
          description: Updated thread
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      summary: Delete thread
      tags: [Threads]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /users/{id}:
    get:
      summary: Get user profile
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile (passwordHash excluded)
        '404':
          description: Not found
    patch:
      summary: Update own profile
      tags: [Users]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                theme:
                  type: string
      responses:
        '200':
          description: Updated user
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /admin/users/{id}:
    patch:
      summary: Update user role/trust (admin only)
      tags: [Admin]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [user, moderator, admin]
                trust:
                  type: string
                  enum: [new, member, trusted, banned]
                isDeleted:
                  type: boolean
                reason:
                  type: string
      responses:
        '200':
          description: Updated user
        '403':
          description: Forbidden
        '404':
          description: Not found
...